Update retailDVDstore
I created a very basic DVD rental program for a final project, and although the final result wasn't as extensive and complete as it could've been this is what I came up with. I plan to revisit this at some point and tackle some of the problems it contain (include movie descriptions, make sure a valid name is included at the beginning and not just any random characters, etc.). I did use a flow chart to give myself an idea of what I desired the program to look like.